# src/openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClawProof — zkML Proof-as-a-Service API",
    "version": "0.1.0",
    "description": "ClawProof generates cryptographic zero-knowledge proofs (zkML) of ML inference. Submit an input to a registered ONNX model, receive a JOLT SNARK proof receipt that any third party can verify. Use this API to prove that a specific model produced a specific output for a given input, without revealing model weights. Proof system: JOLT-Atlas SNARK with Dory/BN254 commitment scheme.",
    "contact": {
      "name": "ClawProof",
      "url": "https://clawproof.onrender.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://clawproof.onrender.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/prove": {
      "post": {
        "operationId": "prove",
        "summary": "Submit input for zkML proof generation",
        "description": "Runs ML inference on the specified model with the provided input, then spawns an asynchronous JOLT SNARK proof generation job. Returns immediately with a receipt ID, the model output (predicted class, label, confidence), and status 'proving'. The proof runs in the background; poll GET /receipt/{id} or provide a webhook_url to be notified when proving completes. Input format depends on the model's input_type: use 'text' for text models, 'fields' for structured_fields models, or 'raw' for raw vector models. Check GET /models to see what each model expects.",
        "tags": ["Proving"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProveRequest"
              },
              "examples": {
                "structured_fields": {
                  "summary": "Structured fields input (authorization model)",
                  "value": {
                    "model_id": "authorization",
                    "input": {
                      "fields": {
                        "budget": 10,
                        "trust": 5,
                        "amount": 3,
                        "category": 1,
                        "velocity": 2,
                        "day": 3,
                        "time": 1,
                        "risk": 0
                      }
                    }
                  }
                },
                "raw": {
                  "summary": "Raw integer vector input",
                  "value": {
                    "model_id": "my_custom_model",
                    "input": {
                      "raw": [1, 0, 0, 3, 7, 0, 0, 2]
                    }
                  }
                },
                "with_webhook": {
                  "summary": "Request with webhook notification",
                  "value": {
                    "model_id": "authorization",
                    "input": {
                      "fields": {
                        "budget": 10,
                        "trust": 5,
                        "amount": 3,
                        "category": 1,
                        "velocity": 2,
                        "day": 3,
                        "time": 1,
                        "risk": 0
                      }
                    },
                    "webhook_url": "https://example.com/hooks/clawproof"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Proof generation started. The receipt_id can be used to poll for proof completion via GET /receipt/{id}. The output contains the model's prediction. Status will be 'proving' until the SNARK proof completes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProveResponse"
                },
                "example": {
                  "receipt_id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                  "receipt_url": "https://clawproof.onrender.com/receipt/a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                  "model_id": "authorization",
                  "output": {
                    "raw_output": [142, -37],
                    "predicted_class": 0,
                    "label": "AUTHORIZED",
                    "confidence": 0.7932960893854749
                  },
                  "status": "proving"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request. Possible causes: missing or wrong input format for the model's input_type, raw vector length mismatch, text exceeds 10,000 characters, webhook_url not HTTPS, or tensor creation failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missing_text": {
                    "summary": "Text input required but not provided",
                    "value": {
                      "error": "Text input required for this model",
                      "hint": "Provide {\"input\": {\"text\": \"...\"}}"
                    }
                  },
                  "missing_fields": {
                    "summary": "Fields input required but not provided",
                    "value": {
                      "error": "Field inputs required for this model",
                      "hint": "Provide {\"input\": {\"fields\": {\"field_name\": value}}}"
                    }
                  },
                  "raw_length_mismatch": {
                    "summary": "Raw input vector has wrong length",
                    "value": {
                      "error": "Raw input length 5 does not match expected 64"
                    }
                  },
                  "webhook_not_https": {
                    "summary": "Webhook URL must use HTTPS",
                    "value": {
                      "error": "webhook_url must use HTTPS",
                      "hint": "Provide a URL starting with https://"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Model not found in the registry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Model not found: nonexistent_model",
                  "hint": "Check GET /models for available model IDs"
                }
              }
            }
          },
          "500": {
            "description": "Internal error during inference or hash computation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Inference failed: ..."
                }
              }
            }
          },
          "503": {
            "description": "Model is registered but still loading (SNARK preprocessing incomplete). Retry after checking GET /health.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Model 'authorization' is still loading. Try again shortly.",
                  "hint": "Check GET /health to see model loading status"
                }
              }
            }
          }
        }
      }
    },
    "/prove/batch": {
      "post": {
        "operationId": "batchProve",
        "summary": "Submit a batch of proof generation requests",
        "description": "Accepts up to 5 proof requests in a single call. Each request follows the same format as POST /prove. All requests are executed sequentially and a combined response is returned. If any individual request fails, the entire batch fails. Use this to generate proofs for multiple model/input combinations in one API call.",
        "tags": ["Proving"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRequest"
              },
              "example": {
                "requests": [
                  {
                    "model_id": "authorization",
                    "input": {
                      "fields": {
                        "budget": 10,
                        "trust": 5,
                        "amount": 3,
                        "category": 1,
                        "velocity": 2,
                        "day": 3,
                        "time": 1,
                        "risk": 0
                      }
                    }
                  },
                  {
                    "model_id": "authorization",
                    "input": {
                      "fields": {
                        "budget": 15,
                        "trust": 1,
                        "amount": 12,
                        "category": 3,
                        "velocity": 5,
                        "day": 0,
                        "time": 3,
                        "risk": 0
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "All batch requests succeeded. Each receipt in the array corresponds positionally to the input request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                },
                "example": {
                  "receipts": [
                    {
                      "receipt_id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "receipt_url": "https://clawproof.onrender.com/receipt/a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "model_id": "authorization",
                      "output": {
                        "raw_output": [142, -37],
                        "predicted_class": 0,
                        "label": "AUTHORIZED",
                        "confidence": 0.7932960893854749
                      },
                      "status": "proving"
                    },
                    {
                      "receipt_id": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
                      "receipt_url": "https://clawproof.onrender.com/receipt/b2c3d4e5-f6a7-8901-bcde-f12345678901",
                      "model_id": "authorization",
                      "output": {
                        "raw_output": [-89, 142],
                        "predicted_class": 1,
                        "label": "DENIED",
                        "confidence": 0.6147186147186147
                      },
                      "status": "proving"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid batch request. The requests array must contain between 1 and 5 items. Individual request validation errors also return 400.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "empty_batch": {
                    "summary": "Empty requests array",
                    "value": {
                      "error": "At least one request is required",
                      "hint": "Provide {\"requests\": [{\"model_id\": \"...\", \"input\": {...}}]}"
                    }
                  },
                  "too_many": {
                    "summary": "Exceeds maximum batch size",
                    "value": {
                      "error": "Maximum 5 requests per batch"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "A referenced model was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error during inference.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A referenced model is still loading.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/receipt/{id}": {
      "get": {
        "operationId": "getReceipt",
        "summary": "Retrieve a proof receipt by ID",
        "description": "Returns the full proof receipt for the given receipt ID. The receipt includes cryptographic hashes (model, input, output, proof), timing information, the inference output, and the current proof status. Supports content negotiation: send Accept: application/json for JSON (default for API clients), or omit for an HTML page. Append ?format=jsonld to get a Schema.org JSON-LD representation suitable for structured data embedding. Poll this endpoint to check when an in-progress proof completes (status transitions from 'proving' to 'verified' or 'failed').",
        "tags": ["Receipts"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The UUID of the proof receipt, returned by POST /prove.",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "description": "Set to 'jsonld' to receive a Schema.org JSON-LD representation of the receipt instead of the default format.",
            "schema": {
              "type": "string",
              "enum": ["jsonld"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The proof receipt. Content type depends on Accept header and format query param.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                },
                "example": {
                  "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                  "model_id": "authorization",
                  "model_name": "Transaction Authorization",
                  "status": "verified",
                  "created_at": "2025-01-15T10:30:00Z",
                  "completed_at": "2025-01-15T10:30:45Z",
                  "model_hash": "sha256:9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08",
                  "input_hash": "sha256:2c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7ae",
                  "output_hash": "sha256:fcde2b2edba56bf408601fb721fe9b5c338d10ee429ea04fae5511b68fbf8fb9",
                  "output": {
                    "raw_output": [142, -37],
                    "predicted_class": 0,
                    "label": "AUTHORIZED",
                    "confidence": 0.7932960893854749
                  },
                  "proof_hash": "sha256:b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944c",
                  "proof_size": 14280,
                  "prove_time_ms": 42350,
                  "verify_time_ms": 120,
                  "error": null
                }
              },
              "application/ld+json": {
                "schema": {
                  "type": "object",
                  "description": "Schema.org JSON-LD representation of the receipt."
                },
                "example": {
                  "@context": "https://schema.org",
                  "@type": "DigitalDocument",
                  "identifier": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                  "name": "zkML Proof Receipt \u2014 Transaction Authorization",
                  "description": "Cryptographic proof of ML inference for model 'authorization'",
                  "dateCreated": "2025-01-15T10:30:00+00:00",
                  "dateModified": "2025-01-15T10:30:45+00:00",
                  "creator": {
                    "@type": "SoftwareApplication",
                    "name": "ClawProof",
                    "url": "https://clawproof.onrender.com"
                  },
                  "about": {
                    "@type": "SoftwareApplication",
                    "name": "Transaction Authorization",
                    "identifier": "authorization"
                  },
                  "status": "verified",
                  "model_hash": "sha256:9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08",
                  "input_hash": "sha256:2c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7ae",
                  "output_hash": "sha256:fcde2b2edba56bf408601fb721fe9b5c338d10ee429ea04fae5511b68fbf8fb9",
                  "proof_hash": "sha256:b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944c",
                  "proof_size": 14280,
                  "prove_time_ms": 42350,
                  "verify_time_ms": 120,
                  "prediction": {
                    "label": "AUTHORIZED",
                    "confidence": 0.7932960893854749,
                    "predicted_class": 0
                  }
                }
              },
              "text/html": {
                "schema": {
                  "type": "string",
                  "description": "Human-readable HTML page showing the receipt details."
                }
              }
            }
          },
          "404": {
            "description": "No receipt found with the given ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Receipt not found",
                  "hint": "Check the receipt ID"
                }
              }
            }
          }
        }
      }
    },
    "/verify": {
      "post": {
        "operationId": "verifyReceipt",
        "summary": "Verify a proof receipt",
        "description": "Checks whether the SNARK proof in the specified receipt is valid. Returns the verification result along with the current receipt status. A receipt with status 'verified' means the cryptographic proof has been generated and independently verified. Status 'proving' means proof generation is still in progress. Status 'failed' means proof generation encountered an error.",
        "tags": ["Verification"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRequest"
              },
              "example": {
                "receipt_id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification result. 'valid' is true only when the proof has been generated and verified successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResponse"
                },
                "examples": {
                  "verified": {
                    "summary": "Proof verified successfully",
                    "value": {
                      "valid": true,
                      "receipt_id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "status": "verified"
                    }
                  },
                  "still_proving": {
                    "summary": "Proof generation still in progress",
                    "value": {
                      "valid": false,
                      "receipt_id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "status": "proving"
                    }
                  },
                  "failed": {
                    "summary": "Proof generation failed",
                    "value": {
                      "valid": false,
                      "receipt_id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "status": "failed"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No receipt found with the given ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Receipt not found",
                  "hint": "Check the receipt_id and try GET /receipt/{id}"
                }
              }
            }
          }
        }
      }
    },
    "/badge/{receipt_id}": {
      "get": {
        "operationId": "getBadge",
        "summary": "Get an SVG badge for a proof receipt",
        "description": "Returns a shields.io-style SVG badge showing the proof status for the given receipt. Embed this in README files, dashboards, or web pages to display real-time proof status. Badge colors: green for 'verified', yellow for 'proving', red for 'failed'. Verified and failed badges are cached for 1 hour; proving badges are not cached.",
        "tags": ["Badges"],
        "parameters": [
          {
            "name": "receipt_id",
            "in": "path",
            "required": true,
            "description": "The UUID of the proof receipt.",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          }
        ],
        "responses": {
          "200": {
            "description": "SVG badge image showing the proof status.",
            "content": {
              "image/svg+xml": {
                "schema": {
                  "type": "string",
                  "description": "SVG markup for the status badge."
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "description": "Caching directive. 'no-cache' for proving receipts, 'public, max-age=3600' for verified/failed.",
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Origin": {
                "description": "Always set to '*' for cross-origin embedding.",
                "schema": {
                  "type": "string",
                  "const": "*"
                }
              }
            }
          },
          "404": {
            "description": "No receipt found with the given ID. Returns plain text 'Receipt not found'.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "Receipt not found"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProveRequest": {
        "type": "object",
        "description": "Request body for generating a zkML proof. Provide the model_id and the appropriate input format for that model's input_type. Optionally include a webhook_url (HTTPS only) to receive a POST notification when the proof completes.",
        "required": ["model_id"],
        "properties": {
          "model_id": {
            "type": "string",
            "description": "The ID of the model to run inference on. Must match a model from GET /models.",
            "example": "authorization"
          },
          "input": {
            "$ref": "#/components/schemas/ProveInput"
          },
          "webhook_url": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "description": "Optional HTTPS URL to receive a POST callback when proof generation completes. Must start with https://.",
            "example": "https://example.com/hooks/clawproof"
          }
        }
      },
...
      "ProveResponse": {
        "type": "object",
        "description": "Response from a successful proof submission. Contains the inference output immediately; the cryptographic proof is generated asynchronously. Poll the receipt_url or use webhook_url to be notified when the proof completes.",
        "required": ["receipt_id", "receipt_url", "model_id", "output", "status"],
        "properties": {
          "receipt_id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for this proof receipt. Use this to check status via GET /receipt/{id} or POST /verify.",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "receipt_url": {
            "type": "string",
            "format": "uri",
            "description": "Direct URL to view or fetch this receipt.",
            "example": "https://clawproof.onrender.com/receipt/a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "model_id": {
            "type": "string",
            "description": "The model that was used for inference.",
            "example": "authorization"
          },
          "output": {
            "$ref": "#/components/schemas/InferenceOutput"
          },
          "status": {
            "type": "string",
            "enum": ["proving"],
            "description": "Always 'proving' in the initial response. The proof is being generated in the background.",
            "example": "proving"
          }
        }
      },
...
    }
  },
...
}